修复user-controller小bug 给将实体类getter和setter调整为链式传值

+++ src/main/java/com/example/musicbox/service/UserService.java
@@ -6,7 +6,7 @@ import com.example.musicbox.entity.User;
 public interface UserService extends IService<User> {
     String login(String username, String password);     // 登录
 
-    int parseToken(String token);   // 解析token
+    long parseToken(String token);   // 解析token
 
     String register(String username, String password);    // 注册
